# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Community Voice Core API",
        "version": "1.0.0"
    },
    "host": "api.community-voice.local",
    "basePath": "/v1",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/postRecordFile": {
            "post": {
                "tags": [
                    "Editor"
                ],
                "summary": "Recorded File Upload",
                "operationId": "postRecordFile",
                "parameters": [
                    {
                        "name": "api-token",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "note",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "duration",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "recorded_file",
                        "in": "formData",
                        "description": "Audio Recorded file Upload ( max : 1GB )",
                        "required": true,
                        "type": "file"
                    },
                    {
                        "name": "image",
                        "in": "formData",
                        "description": "Audio Photo Upload (max : 2MB )",
                        "required": false,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation !"
                    },
                    "204": {
                        "description": "No Content !"
                    },
                    "400": {
                        "description": "Invalid request !",
                        "examples": {
                            "application/json": {
                                "status": 400,
                                "message": "Custome Error Message!"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error !",
                        "examples": {
                            "application/json": {
                                "status": 500,
                                "message": "Something wrong!"
                            }
                        }
                    }
                }
            }
        },
        "/getUploadedRecord": {
            "get": {
                "tags": [
                    "Editor"
                ],
                "summary": "Get Uploaded Records",
                "operationId": "getUploadedRecord",
                "parameters": [
                    {
                        "name": "api-token",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation !",
                        "examples": {
                            "application/json": {
                                "data": [
                                    {
                                        "id": "2384c529-f0c4-445e-9f51-f5997b78e2e4",
                                        "title": "Est earum eveniet ullam minima.",
                                        "duration": 2925000,
                                        "created_at": "December 6 2019, 4:28 am",
                                        "image": null,
                                        "audio": null,
                                        "note": "Enim sunt est nemo quis voluptas et.",
                                        "station_title": "ut"
                                    }
                                ]
                            }
                        }
                    },
                    "204": {
                        "description": "No Content !"
                    },
                    "400": {
                        "description": "Invalid request !",
                        "examples": {
                            "application/json": {
                                "status": 400,
                                "message": "Custome Error Message!"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error !",
                        "examples": {
                            "application/json": {
                                "status": 500,
                                "message": "Something wrong!"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}